p3 woche 4 fertig

+++ src/p2/P2_main.java
@@ -4,9 +4,9 @@ public class P2_main {
 
 	public static void main(String[] args) {
 		
-		int a = 3, b = 5, c = 6;
+		double a = 3, b = 5, c = 6;
 		double x = 1.5, y = 2.3;
-		int intResult;
+		double intResult;
 		double doubleResult;
 		
 		intResult = c - a * 3;
